+++ packages/extension/components/icons/Ellipsis.vue
@@ -1,0 +1,8 @@
+<template>
+  <svg class="tp-icon" width="24" height="24" fill="none" viewBox="0 0 24 24">
+    <path
+      fill="currentColor"
+      d="M7 13a1 1 0 1 0 0-2 1 1 0 0 0 0 2m5 0a1 1 0 1 0 0-2 1 1 0 0 0 0 2m5 0a1 1 0 1 0 0-2 1 1 0 0 0 0 2"
+    ></path>
+  </svg>
+</template>